requirements.txt
PyPDF2
langchain
langchain_community
chainlit
langchain_groq
python-dotenv
chromadb